change Icesat2data class to Query class (#102)
* change icesat2data class to query class. Update docs and examples accordingly. Add deprecation warning for icesat2data.
