+++ doc/source/user_guide/documentation/components.rst
@@ -1,4 +1,4 @@
-Icesat2Data Components
+Query Components
 ======================
