+++ doc/source/user_guide/documentation/icepyx_architecture_user.wsd
@@ -1,7 +1,7 @@
 @startuml
-title **icepyx Class Diagram** <img: ../_static/icepyx_v2_oval_tiny-uml.PNG>
+title **icepyx Class Diagram** <img: ../../_static/icepyx_v2_oval_tiny-uml.PNG>
 
-class Icesat2Data {
+class Query {
     reqparams
     extent_type
     subsetparams
@@ -157,25 +157,25 @@ package "Format for API (apifmt)" {
 }
 
 
-Icesat2Data *-left- Granules 
-Icesat2Data *-right- Variables 
+Query *-left- Granules 
+Query *-right- Variables 
 
-''Icesat2Data -right-> about_dataset : is2ref.about_dataset
-''Icesat2Data -right-> _get_custom_options : is2ref._get_custom_options
-''Icesat2Data -right-> _validate_dataset : is2ref._validate_dataset
-Icesat2Data -right-> "ICESat-2 Reference (is2ref)"
+''Query -right-> about_dataset : is2ref.about_dataset
+''Query -right-> _get_custom_options : is2ref._get_custom_options
+''Query -right-> _validate_dataset : is2ref._validate_dataset
+Query -right-> "ICESat-2 Reference (is2ref)"
 
-Icesat2Data -right-> "Validation (val)"
-''Icesat2Data --> dset_version : val.dset_version
-''Icesat2Data --> spatial : val.spatial
-''Icesat2Data --> temporal: val.temporal
+Query -right-> "Validation (val)"
+''Query --> dset_version : val.dset_version
+''Query --> spatial : val.spatial
+''Query --> temporal: val.temporal
 
-Icesat2Data -left-> "Format for API (apifmt)"
-''Icesat2Data --> build_CMR_params: apifmt.build_CMR_params
-''Icesat2Data -->  build_reqconfig_params : apifmt.build_reqconfig_params
-'' Icesat2Data -left-> APIformatting
+Query -left-> "Format for API (apifmt)"
+''Query --> build_CMR_params: apifmt.build_CMR_params
+''Query -->  build_reqconfig_params : apifmt.build_reqconfig_params
+'' Query -left-> APIformatting
 
-Icesat2Data .down. Earthdata
+Query .down. Earthdata
 
 ' Variables -down-> _get_custom_options
