+++ doc/source/user_guide/documentation/icesat2data.rst
@@ -1,7 +1,7 @@
-Icesat2Data Class
+Query Class
 =================
 
-.. currentmodule:: icepyx.icesat2data
+.. currentmodule:: icepyx.query
 
 
 Constructor
@@ -10,7 +10,7 @@ Constructor
 .. autosummary::
    :toctree: ../_icepyx/
 
-   Icesat2Data
+   Query
 
 
 Attributes
@@ -19,18 +19,18 @@ Attributes
 .. autosummary::
    :toctree: ../_icepyx/
    
-   Icesat2Data.CMRparams
-   Icesat2Data.dataset
-   Icesat2Data.dataset_version
-   Icesat2Data.dates
-   Icesat2Data.end_time
-   Icesat2Data.file_vars
-   Icesat2Data.granules
-   Icesat2Data.order_vars
-   Icesat2Data.reqparams
-   Icesat2Data.spatial_extent
-   Icesat2Data.subsetparams
-   Icesat2Data.start_time
+   Query.CMRparams
+   Query.dataset
+   Query.dataset_version
+   Query.dates
+   Query.end_time
+   Query.file_vars
+   Query.granules
+   Query.order_vars
+   Query.reqparams
+   Query.spatial_extent
+   Query.subsetparams
+   Query.start_time
 
 
 Methods
@@ -39,12 +39,12 @@ Methods
 .. autosummary::
    :toctree: ../_icepyx/
    
-   Icesat2Data.avail_granules
-   Icesat2Data.dataset_all_info
-   Icesat2Data.dataset_summary_info
-   Icesat2Data.download_granules
-   Icesat2Data.earthdata_login
-   Icesat2Data.latest_version
-   Icesat2Data.order_granules
-   Icesat2Data.show_custom_options
-   Icesat2Data.visualize_spatial_extent
+   Query.avail_granules
+   Query.dataset_all_info
+   Query.dataset_summary_info
+   Query.download_granules
+   Query.earthdata_login
+   Query.latest_version
+   Query.order_granules
+   Query.show_custom_options
+   Query.visualize_spatial_extent
