+++ icepyx/tests/test_granules.py
@@ -1,14 +1,14 @@
 import pytest
 import warnings
 
-from icepyx.core import icesat2data as ipd
+from icepyx.core import query as ipd
 from icepyx.core import granules as granules
 from icepyx.core.granules import Granules as Granules
 
 
 # @pytest.fixture
 # def reg_a():
-#     return ipd.Icesat2Data('ATL06',[-55, 68, -48, 71],['2019-02-22','2019-02-28'])
+#     return ipd.Query('ATL06',[-55, 68, -48, 71],['2019-02-22','2019-02-28'])
 
 # #@patch('my_module.__get_input', return_value='y')
 
@@ -19,7 +19,7 @@ from icepyx.core.granules import Granules as Granules
 
 # DevNote: clearly there's a better way that doesn't make the function so long... what is it?
 def test_granules_info():
-    # reg_a = ipd.Icesat2Data('ATL06', [-55, 68, -48, 71], ['2019-02-20','2019-02-24'], version='3')
+    # reg_a = ipd.Query('ATL06', [-55, 68, -48, 71], ['2019-02-20','2019-02-24'], version='3')
     # granules = reg_a.granules.avail
     grans = [
         {
@@ -585,15 +585,16 @@ def test_granules_info():
 def test_no_granules_in_search_results():
     ermsg = "Your search returned no results; try different search parameters"
     with pytest.raises(AssertionError, match=ermsg):
-        ipd.Icesat2Data(
+        ipd.Query(
             "ATL06", [-55, 68, -48, 71], ["2019-02-20", "2019-02-20"], version="2"
         ).avail_granules()
 
 
 def test_correct_granule_list_returned():
-    reg_a = ipd.Icesat2Data(
+    reg_a = ipd.Query(
         "ATL06", [-55, 68, -48, 71], ["2019-02-20", "2019-02-28"], version="2"
     )
+
     reg_a.avail_granules()
     obs_grans = [gran["producer_granule_id"] for gran in reg_a.granules.avail]
     exp_grans = [
